add `desired_deps` attr to `apply_fixes`
This attr is a string, which must be a valid argument to `bazel cquery`,
that specifies a pattern that `deps` should expand to. This can be used
to keep the `deps` attr in sync e.g. a new test target is defined but
the `apply_fixes` rule is not updated.

`deps` sync is not reliable run and may necessitate use of `bazel
clean` to force a re-run.