manually add dependencies to `tidy_test` rule (#15)
Update `tidy_test` with explicitly specified dependencies on the
external example directory and files within `rules_clang_tidy`.

The example directory is excluded from the workspace with `.bazelignore`
and is brought in with `local_repository`. Changes to any file in the
example directory will trigger a retest by use of
`local_workspace_status` to generate a status file.

Files within `rules_clang_tidy` are added explicitly as they are not
direct dependencies of the test scripts. Instead they are transitive
dependencies of the targets within the example directories.

Change-Id: I39f0b5336bd7bf9c5700e52400a91dfce2fa5fc7