always generate fixes in ClangTidy aspect
Always generate the exported fixes when executing a ClangTidy aspect. By
using the `check` aspect in the `apply_fixes` rule, this rule can use
fixes stored in the action cache if available.

As `apply_fixes` now uses the `check` aspect, this commit removes the
`export_fixes` aspect as it is no longer used.

`--remote_download_toplevel` is suggested as an option in `.bazelrc` so
that aspect outputs may be downloaded from a remote cache if present.
This is primarily for use with the `apply-fixes` script which does not
define fixes as inputs.

Change-Id: I3c277a9bad789229b4138d7993d0a609fff0925d